+++ src/imio/esign/browser/views.py
@@ -4,7 +4,6 @@ from AccessControl import Unauthorized
 from datetime import datetime
 from datetime import timedelta
 from imio.esign import _
-from imio.esign import ESIGN_CREDENTIALS
 from imio.esign import ESIGN_ROOT_URL
 from imio.esign import manage_session_perm
 from imio.esign.browser.table import external_session_link
@@ -12,6 +11,7 @@ from imio.esign.browser.table import SessionsTable
 from imio.esign.config import get_registry_enabled
 from imio.esign.utils import create_external_session
 from imio.esign.utils import get_session_annotation
+from imio.esign.utils import get_state_description
 from imio.esign.utils import remove_session
 from imio.helpers.content import uuidToObject
 from imio.helpers.security import separate_fullname
@@ -146,7 +146,6 @@ class ExternalSessionCreateView(BrowserView):
             return self.context.absolute_url() + "/@@parapheo"
         resp = create_external_session(
             int(session_id),
-            b64_cred=ESIGN_CREDENTIALS,
             esign_root_url=ESIGN_ROOT_URL,
         )
         if resp == "_session_not_found_":
@@ -235,6 +234,9 @@ class FacetedSessionInfoViewlet(ViewletBase):
     def collapsible_content_css_default(self):
         """Default CSS class to apply on the collapsible."""
         return "collapsible-content"
+
+    def get_state_description(self, state):
+        return translate(get_state_description(state), context=self.request, domain="imio.esign")
 
 
 class ItemSessionInfoViewlet(FacetedSessionInfoViewlet):