+++ src/imio/esign/utils.py
@@ -2,10 +2,10 @@
 from datetime import datetime
 from datetime import timedelta
 from imio.esign import _tr as _
-from imio.esign import ESIGN_CREDENTIALS
 from imio.esign import ESIGN_ROOT_URL
 from imio.esign import logger
 from imio.esign.config import get_registry_file_url
+from imio.esign.config import get_registry_max_session_size
 from imio.esign.config import get_registry_seal_code
 from imio.esign.config import get_registry_seal_email
 from imio.esign.config import get_registry_sign_code
@@ -14,7 +14,7 @@ from imio.esign.interfaces import IContextUidProvider
 from imio.helpers.content import uuidsToObjects
 from imio.helpers.content import uuidToObject
 from imio.helpers.transmogrifier import get_correct_id
-from imio.pyutils.system import post_request
+from imio.helpers.ws import get_auth_token
 from imio.pyutils.utils import shortuid_encode_id
 from os import path
 from persistent.list import PersistentList
@@ -22,15 +22,30 @@ from persistent.mapping import PersistentMapping
 from plone import api
 from zope.annotation import IAnnotations
 from zope.component import getAdapter
+from zope.i18n import translate
 
 import json
+import requests
 
 
 SESSION_URL = "imio/esign/v1/luxtrust/sessions"
 
 
+def get_filesize(uid):
+    """Get the file size of an annex.
+
+    :param uid: The UID of the annex.
+    :return: The file size in bytes, or 0 if not found.
+    """
+    annex = uuidToObject(uuid=uid, unrestricted=True)
+    if not annex:
+        logger.error("Annex with UID %s not found.", uid)
+        return 0
+    return getattr(annex.__parent__, "categorized_elements", {}).get(uid, {}).get("filesize", annex.file.size)
+
+
 def add_files_to_session(
-    signers, files_uids, seal=None, acroform=True, session_id=None, title=None, discriminators=(), watchers=(),
+    signers, files_uids, seal=None, acroform=True, session_id=None, title="", discriminators=(), watchers=(),
     create_session_custom_data=None
 ):
     """Add files to a session with the given signers.
@@ -47,6 +62,7 @@ def add_files_to_session(
     :return: session_id, session
     """
     annot = get_session_annotation()
+    size = sum(get_filesize(uid) for uid in files_uids)
     if session_id is not None:
         if session_id not in annot["sessions"]:
             logger.error("Session with id %s not found in esign annotations.", session_id)
@@ -54,13 +70,14 @@ def add_files_to_session(
         else:
             session = annot["sessions"][session_id]
     else:
-        session_id, session = discriminate_sessions(signers, seal, acroform, discriminators=discriminators)
+        session_id, session = discriminate_sessions(signers, seal, acroform, discriminators=discriminators, size=size)
     if not session:
         session_id, session = create_session(
-            signers, seal, acroform=acroform, title=title or "", annot=annot, discriminators=discriminators,
+            signers, seal, acroform=acroform, title=title, annot=annot, discriminators=discriminators,
             watchers=watchers,
             create_session_custom_data=create_session_custom_data,
         )
+    session["size"] = session.get("size", 0) + size
     existing_files = [path.splitext(f["filename"])[0] for f in session["files"]]
     for uid in files_uids:
         annex = uuidToObject(uuid=uid, unrestricted=True)
@@ -94,11 +111,10 @@ def add_files_to_session(
     return session_id, session
 
 
-def create_external_session(session_id, b64_cred=None, esign_root_url=None):
+def create_external_session(session_id, esign_root_url=None):
     """Create a session with the given signers and files.
 
     :param session_id: internal session id
-    :param b64_cred: base64 encoded credentials for authentication
     :param esign_root_url: the root URL for the e-sign service, if not provided it will use the default ESIGN_ROOT_URL
     :return: session information
     """
@@ -153,17 +169,22 @@ def create_external_session(session_id, b64_cred=None, esign_root_url=None):
             "sealCode": seal_code,
         }
 
+    # files_payload = {filename: file_content for z, filename, file_content, uid in files}
     files_payload = [("files", (filename, file_content)) for z, filename, file_content, uid in files]
 
     # Headers avec autorisation
-    headers = {"accept": "application/json"}
-    b64_cred = b64_cred or ESIGN_CREDENTIALS
-    if b64_cred:
-        headers["Authorization"] = "Basic {}".format(b64_cred)
+    headers = {
+        "accept": "application/json",
+        "Authorization": "Bearer %s" % get_auth_token(),
+    }
 
     logger.info(data_payload)
-    ret = post_request(
-        session_url, data={"data": json.dumps(data_payload, default=vars)}, headers=headers, files=files_payload
+    ret = requests.post(
+        session_url,
+        headers=headers,
+        data={"data": json.dumps(data_payload, default=vars)},
+        files=files_payload,
+        timeout=10,
     )
     if ret.status_code == 200:
         session["state"] = "sent"
@@ -208,7 +229,7 @@ def create_session(signers, seal=False, acroform=True, title=None, annot=None, d
             ]
         ),
         "watchers": PersistentList(watchers),
-        "state": "draft",  # draft, sent, errored, to_sign, signed, refused, returned, finalized
+        "state": "draft",
         "title": title,
         "returns": PersistentList(),
     })
@@ -218,19 +239,21 @@ def create_session(signers, seal=False, acroform=True, title=None, annot=None, d
     return session_id, sessions[session_id]
 
 
-def discriminate_sessions(signers, seal, acroform, discriminators=(), annot=None):
+def discriminate_sessions(signers, seal, acroform, discriminators=(), annot=None, size=0):
     """Discriminate sessions based on seal value and signers in the same order.
 
     :param signers: a list of signers, each is a tuple with userid and email
     :param seal: seal boolean
     :param acroform: boolean value indicating if acroform is used
     :param discriminators: optional list of string discriminators
+    :param size: size in bytes of the files to be added to the session
     :param annot: esign annotation, if not provided it will be fetched
     :return: session id and session if found, or (None, None) if no session found
     """
     if not annot:
         annot = get_session_annotation()
     sessions = annot.get("sessions", {})
+    max_session_size = get_registry_max_session_size() * 1024**2
 
     for session_id, session in sessions.items():
         if session["state"] != "draft":
@@ -241,6 +264,8 @@ def discriminate_sessions(signers, seal, acroform, discriminators=(), annot=None
             continue
         session_signers = session.get("signers", [])
         if len(signers) != len(session_signers):
+            continue
+        if size + session.get("size", 0) > max_session_size:
             continue
 
         if set(discriminators) != set(session.get("discriminators", ())):
@@ -352,6 +377,7 @@ def remove_files_from_session(files_uids):
             logger.error("Session %s not found", session_id)
             continue
         session = sessions[session_id]
+        session["size"] = max(0, session.get("size", 0) - get_filesize(uid))
         i = 0
         context_uid = None
         for j, dic in enumerate(session["files"]):
@@ -442,3 +468,55 @@ def get_suid_from_uuid(uid):
     :return: short UID
     """
     return shortuid_encode_id(uid, separator="-", block_size=5)
+
+
+def get_state_description(state):
+    """
+    Get a human readable description for a given session state.
+
+                       ┌─────────┐
+                       │  draft  │
+                       └────┬────┘
+                            │
+                 (session sent to Paraphéo)
+                            │
+                            ▼
+                       ┌─────────┐
+             ┌─────────│  sent   │─────────┐
+             │         └────┬────┘         │
+             │              │              │
+    (error occurred)     (ready)   (signer refused)
+             │              │              │
+             ▼              ▼              ▼
+        ┌─────────┐    ┌─────────┐    ┌─────────┐
+        │ errored │    │ to_sign │    │ refused │
+        └─────────┘    └────┬────┘    └─────────┘
+                            │
+                   (documents signed)
+                            │
+                            ├──────────────┐
+                            │              │
+            (sent back successfully)  (send back failed)
+                            │              │
+                            ▼              │
+                      ┌──────────┐         │
+                      │ returned │         │
+                      └─────┬────┘         │
+                            │              │
+                 (documents received)      │
+                            │              │
+                            ▼              ▼
+                      ┌───────────┐   ┌─────────┐
+                      │ finalized │   │ signed  │
+                      └───────────┘   └─────────┘
+    """
+    return {
+        'draft': u'The session is getting ready to be sent to Paraphéo by a signing manager.',
+        'sent': u'The session has been sent to Paraphéo.',
+        'errored': u'The session encountered an error during its processing.',
+        'to_sign': u'The session is ready to be signed in Paraphéo.',
+        'signed': u'The session is finished but signed documents couldn\'t be sent back to the application.',
+        'refused': u'The session has been cancelled because a signer refused a document.',
+        'returned': u'The session is finished and signed documents are on the way back to the application.',
+        'finalized': u'The session is finished and signed documents have been sent back to the application.',
+    }.get(state, "")