+++ CHANGES.rst
@@ -7,6 +7,14 @@ Changelog
 
 - Replaced external_session_link p by span.
   [sgeulette]
+- Switched basic auth to jwt.
+  [chris-adam]
+- Added files size session discriminator
+  [chris-adam, sgeulette]
+- Added title attribute on session state for better description
+  [chris-adam]
+- Raise Unauthorized if accessing @@parapheo and not available
+  [gbastien]
 - Added action to remove a single item from session.
   [chris-adam]